Issue with setting a default group
I have been unsuccessful in setting users to a default group. Chasing the code down, I found that in the UserManager file on line 214 the conditional seems to get skipped in some circumstance - moving the updateUser() method outside of this conditional seems to have fixed the issue and groups are assigned upon creating users.